requirements.txt
django
gunicorn
django-heroku
psycopg2
dj_database_url
django-taggit
Pillow
django-shopping-cart